moved all spec files to same folder location
