+++ seedDatabase.js
@@ -1,5 +1,5 @@
-const dbHelpers = require('./dbHelpers');
-const sampleData = require('../data/sampleData');
+const dbHelpers = require('./db/dbHelpers');
+const sampleData = require('./data/sampleData');
 
 dbHelpers.save({ data: sampleData, model: dbHelpers.Restaurant }, (result) => {
   if (result) {
